Move all platforms into its own folder
