+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules
 vendor
+.idea
 *.swp
 .DS_Store
 Thumbs.db
