+++ src/Migration/Platforms/SEOPress.php
@@ -1,5 +1,5 @@
 <?php
-namespace SlimSEO\Migration;
+namespace SlimSEO\Migration\Platforms;
 
 class SEOPress extends Replacer {
 	private $context;
