+++ src/Migration/ReplacerFactory.php
@@ -5,19 +5,19 @@ class ReplacerFactory {
 	public static function make( $platform ) {
 		switch ( $platform ) {
 			case 'yoast':
-				return new Yoast;
+				return new Platforms\Yoast;
 			case 'aioseo':
-				return new AIOSEO;
+				return new Platforms\AIOSEO;
 			case 'seo-framework':
-				return new SEOFramework;
+				return new Platforms\SEOFramework;
 			case 'rank-math':
-				return new RankMath;
+				return new Platforms\RankMath;
 			case 'seopress':
-				return new SEOPress;
+				return new Platforms\SEOPress;
 			case 'redirection':
-				return new Redirection;
+				return new Platforms\Redirection;
 			case '301-redirects':
-				return new Redirects301;
+				return new Platforms\Redirects301;
 		}
 	}
 }
